+++ marathon/provider.go
@@ -8,7 +8,6 @@ import (
 	"github.com/gambol99/go-marathon"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/terraform"
-	"strings"
 )
 
 // Provider is the provider for terraform
@@ -71,10 +70,8 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 	marathonConfig := marathon.NewDefaultConfig()
 
 	marathonConfig.URL = d.Get("url").(string)
-	if ! strings.HasPrefix(marathonConfig.URL, "https") {
-		marathonConfig.HTTPClient = &http.Client{
-			Timeout: time.Duration(d.Get("request_timeout").(int)) * time.Second,
-		}
+	marathonConfig.HTTPClient = &http.Client{
+		Timeout: time.Duration(d.Get("request_timeout").(int)) * time.Second,
 	}
 
 	marathonConfig.HTTPBasicAuthUser = d.Get("basic_auth_user").(string)
@@ -85,13 +82,11 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 	}
 	marathonConfig.EventsTransport = marathon.EventsTransportSSE
 
-	//if ! strings.HasPrefix(marathonConfig.URL, "https") {
-	//	marathonConfig.HTTPClient = &http.Client{
-	//		Timeout: (time.Duration(d.Get("deployment_timeout").(int)) * time.Second),
-	//	}
-	//}
+	marathonConfig.HTTPClient = &http.Client{
+		Timeout: (time.Duration(d.Get("deployment_timeout").(int)) * time.Second),
+	}
 
-	marathonConfig.HTTPClient = &http.Client{}
+	marathonConfig.HTTPSSEClient = &http.Client{}
 
 	config := config{
 		config: marathonConfig,